libs/damap/src/lib/components/dmp/specify-data/specify-data-instruction/specify-data-instruction.component.ts
import { Component, EventEmitter, Output } from '@angular/core';

@Component({
  selector: 'app-specify-data-instruction',
  templateUrl: './specify-data-instruction.component.html',
})
export class SpecifyDataInstructionComponent {
  @Output() selectionChange = new EventEmitter<
    'primaryView' | 'secondaryView'
  >();

  selectedView: 'primaryView' | 'secondaryView' = 'primaryView';

  onSelectionChange(view: 'primaryView' | 'secondaryView'): void {
    this.selectedView = view;
    this.emitSelection(view);
  }

  private emitSelection(view: 'primaryView' | 'secondaryView'): void {
    this.selectionChange.emit(view);
  }
}
